+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@ Version 3.17.0 (August 3, 2021)
 
 **User-visible changes:**
 
-Java records are type-checked.
+Java records are type-checked.  Thanks to Neil Brown.
 
 `-Ainfer` can now infer postcondition annotations that reference formal parameters
 (e.g. `"#1"`, `"#2"`) and the receiver (`"this"`).