+++ docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 3.17.0 (August 3, 2021)
+Version 3.17.0 (August 2, 2021)
 -------------------------------
 
 **User-visible changes:**
@@ -27,7 +27,7 @@ The following methods no longer take a `fieldValues` parameter:
  * `CFAbstractAnalysis` constructors
 
 **Closed issues:**
-
+#4685, #4689, #4785, #4805, #4806, #4815, #4829, #4849.
 
 Version 3.16.0 (July 13, 2021)
 ------------------------------